feat: add full stack web app
This change implements a new Full Stack Next.js web application, combining two core modern-day powerful technologies (React.js and FastAPI).

The result of the implementation is a fully functioning interactive frontend client that is accessible from the web browser, allowing the user to preview and upload their resumes and multiple job details (link and descriptions) to the backend API, to then be able to take in the user-submitted data, and return out (currently mocked) results, such as vector scores, common words, and suggestions against multiple jobs.

At the moment, the API returns a mocked response, but the API response model is provisionally defined, and so this should make it simpler to get started with integrating new and existing Python scripts to return the results in a structured data format for the frontend client to automatically parse and render back out to the user.