chore: error handling for missing keys config file
Add error handling on the backend api for when `similarity/config.yaml` does not exist, to which it responds with a 404 not found response.

As a result, also added error handling on the frontend ui code to handle the above 404 backend api response edge case to not render the Service Keys ui section.