chore: update readme notify webapp fake data
Add an important note to the README to notify community that the full stack web app currently returns mocked / faked data for the results (e.g. Vector Scores, Common Keyword Matches, Suggestions), and that it will be further developed out and implemented with real results in a later future release.